requirements.txt
streamlit==1.22.0
selenium==4.9.1
beautifulsoup4==4.12.2
pandas==2.0.1
requests==2.30.0
python-dotenv==1.0.0
webdriver-manager==3.8.6
typing-extensions==4.5.0
pyyaml==6.0
retry==0.9.2
pytest==7.3.1
pytest-cov==4.0.0
mypy==1.3.0
flake8==6.0.0
black==23.3.0